Consolidate UAV ROS test stubs
