+++ controls/sae_2025_ws/src/uav/test/test_auto_launch.py
@@ -1,234 +1,16 @@
 from __future__ import annotations
 
-import importlib
 import importlib.util
 from pathlib import Path
-import sys
 from types import SimpleNamespace
-import types
 from typing import Any, cast
 
 import pytest
 
-
-def _import_module_if_available(name: str):
-    try:
-        return importlib.import_module(name)
-    except ModuleNotFoundError:
-        return None
+from uav.test_support.ros_stubs import install_auto_launch_import_stubs
 
 
-std_msgs: Any
-std_msgs_msg: Any
-if "std_msgs" not in sys.modules:
-    std_msgs = _import_module_if_available("std_msgs")
-else:
-    std_msgs = sys.modules["std_msgs"]
-if std_msgs is None:
-    std_msgs = types.ModuleType("std_msgs")
-    std_msgs_msg = types.ModuleType("std_msgs.msg")
-    std_msgs_msg.Empty = type("Empty", (), {})
-    std_msgs.msg = std_msgs_msg
-    sys.modules.update({"std_msgs": std_msgs, "std_msgs.msg": std_msgs_msg})
-
-ament_index_python: Any = sys.modules.get("ament_index_python")
-if ament_index_python is None:
-    ament_index_python = _import_module_if_available("ament_index_python")
-if ament_index_python is None:
-    ament_index_python = types.ModuleType("ament_index_python")
-    sys.modules["ament_index_python"] = ament_index_python
-
-ament_index_packages: Any = sys.modules.get("ament_index_python.packages")
-if ament_index_packages is None:
-    ament_index_packages = _import_module_if_available("ament_index_python.packages")
-if ament_index_packages is None:
-    ament_index_packages = types.ModuleType("ament_index_python.packages")
-    sys.modules["ament_index_python.packages"] = ament_index_packages
-if not hasattr(ament_index_packages, "PackageNotFoundError"):
-
-    class PackageNotFoundError(Exception):
-        pass
-
-    ament_index_packages.PackageNotFoundError = PackageNotFoundError
-if not hasattr(ament_index_packages, "get_package_share_directory"):
-    ament_index_packages.get_package_share_directory = lambda _name: str(
-        Path(__file__).resolve().parents[1]
-    )
-setattr(sys.modules["ament_index_python"], "packages", ament_index_packages)
-
-std_srvs: Any
-std_srvs_srv: Any
-if "std_srvs" not in sys.modules:
-    std_srvs = _import_module_if_available("std_srvs")
-else:
-    std_srvs = sys.modules["std_srvs"]
-if std_srvs is None:
-    std_srvs = types.ModuleType("std_srvs")
-    std_srvs_srv = types.ModuleType("std_srvs.srv")
-
-    class Trigger:
-        Request = type("Request", (), {})
-        Response = type("Response", (), {})
-
-    std_srvs_srv.Trigger = Trigger
-    std_srvs.srv = std_srvs_srv
-    sys.modules.update({"std_srvs": std_srvs, "std_srvs.srv": std_srvs_srv})
-
-launch_module: Any = sys.modules.get("launch")
-if launch_module is None:
-    launch_module = _import_module_if_available("launch")
-if launch_module is None:
-    launch_module = types.ModuleType("launch")
-    sys.modules["launch"] = launch_module
-if not hasattr(launch_module, "LaunchDescription"):
-    launch_module.LaunchDescription = type("LaunchDescription", (), {})
-
-launch_actions: Any = sys.modules.get("launch.actions")
-if launch_actions is None:
-    launch_actions = _import_module_if_available("launch.actions")
-if launch_actions is None:
-    launch_actions = types.ModuleType("launch.actions")
-    sys.modules["launch.actions"] = launch_actions
-for name in (
-    "DeclareLaunchArgument",
-    "ExecuteProcess",
-    "IncludeLaunchDescription",
-    "OpaqueFunction",
-):
-    if not hasattr(launch_actions, name):
-        setattr(launch_actions, name, type(name, (), {}))
-
-launch_sources: Any = sys.modules.get("launch.launch_description_sources")
-if launch_sources is None:
-    launch_sources = _import_module_if_available("launch.launch_description_sources")
-if launch_sources is None:
-    launch_sources = types.ModuleType("launch.launch_description_sources")
-    sys.modules["launch.launch_description_sources"] = launch_sources
-if not hasattr(launch_sources, "PythonLaunchDescriptionSource"):
-    launch_sources.PythonLaunchDescriptionSource = type(
-        "PythonLaunchDescriptionSource", (), {}
-    )
-
-launch_logging: Any = sys.modules.get("launch.logging")
-if launch_logging is None:
-    launch_logging = _import_module_if_available("launch.logging")
-if launch_logging is None:
-    launch_logging = types.ModuleType("launch.logging")
-    sys.modules["launch.logging"] = launch_logging
-if not hasattr(launch_logging, "get_logger"):
-    launch_logging.get_logger = lambda *_args, **_kwargs: SimpleNamespace(
-        warning=lambda *_a, **_k: None,
-        warn=lambda *_a, **_k: None,
-        info=lambda *_a, **_k: None,
-    )
-
-launch_substitutions: Any = sys.modules.get("launch.substitutions")
-if launch_substitutions is None:
-    launch_substitutions = _import_module_if_available("launch.substitutions")
-if launch_substitutions is None:
-    launch_substitutions = types.ModuleType("launch.substitutions")
-    sys.modules["launch.substitutions"] = launch_substitutions
-if not hasattr(launch_substitutions, "LaunchConfiguration"):
-    launch_substitutions.LaunchConfiguration = type("LaunchConfiguration", (), {})
-
-rclpy: Any = sys.modules.get("rclpy")
-if rclpy is None:
-    rclpy = _import_module_if_available("rclpy")
-if rclpy is None:
-    rclpy = types.ModuleType("rclpy")
-    sys.modules["rclpy"] = rclpy
-if not hasattr(rclpy, "init"):
-    rclpy.init = lambda *args, **kwargs: None
-if not hasattr(rclpy, "shutdown"):
-    rclpy.shutdown = lambda: None
-if not hasattr(rclpy, "ok"):
-    rclpy.ok = lambda: True
-
-node_mod: Any = sys.modules.get("rclpy.node")
-if node_mod is None:
-    node_mod = _import_module_if_available("rclpy.node")
-if node_mod is None:
-    node_mod = types.ModuleType("rclpy.node")
-    sys.modules["rclpy.node"] = node_mod
-if not hasattr(node_mod, "Node"):
-
-    class Node:
-        def __init__(self, *_args, **_kwargs) -> None:
-            pass
-
-    node_mod.Node = Node
-
-executors_mod: Any = sys.modules.get("rclpy.executors")
-if executors_mod is None:
-    executors_mod = _import_module_if_available("rclpy.executors")
-if executors_mod is None:
-    executors_mod = types.ModuleType("rclpy.executors")
-    sys.modules["rclpy.executors"] = executors_mod
-if not hasattr(executors_mod, "ExternalShutdownException"):
-
-    class ExternalShutdownException(Exception):
-        pass
-
-    executors_mod.ExternalShutdownException = ExternalShutdownException
-
-clock_mod: Any = sys.modules.get("rclpy.clock")
-if clock_mod is None:
-    clock_mod = _import_module_if_available("rclpy.clock")
-if clock_mod is None:
-    clock_mod = types.ModuleType("rclpy.clock")
-    sys.modules["rclpy.clock"] = clock_mod
-if not hasattr(clock_mod, "Clock"):
-    clock_mod.Clock = type("Clock", (), {})
-
-parameter_mod: Any = sys.modules.get("rclpy.parameter")
-if parameter_mod is None:
-    parameter_mod = _import_module_if_available("rclpy.parameter")
-if parameter_mod is None:
-    parameter_mod = types.ModuleType("rclpy.parameter")
-    sys.modules["rclpy.parameter"] = parameter_mod
-if not hasattr(parameter_mod, "Parameter"):
-    parameter_mod.Parameter = type("Parameter", (), {})
-
-validate_namespace_mod: Any = sys.modules.get("rclpy.validate_namespace")
-if validate_namespace_mod is None:
-    validate_namespace_mod = _import_module_if_available("rclpy.validate_namespace")
-if validate_namespace_mod is None:
-    validate_namespace_mod = types.ModuleType("rclpy.validate_namespace")
-    sys.modules["rclpy.validate_namespace"] = validate_namespace_mod
-if not hasattr(validate_namespace_mod, "validate_namespace"):
-    validate_namespace_mod.validate_namespace = lambda namespace: None
-
-validate_node_name_mod: Any = sys.modules.get("rclpy.validate_node_name")
-if validate_node_name_mod is None:
-    validate_node_name_mod = _import_module_if_available("rclpy.validate_node_name")
-if validate_node_name_mod is None:
-    validate_node_name_mod = types.ModuleType("rclpy.validate_node_name")
-    sys.modules["rclpy.validate_node_name"] = validate_node_name_mod
-if not hasattr(validate_node_name_mod, "validate_node_name"):
-    validate_node_name_mod.validate_node_name = lambda node_name: None
-
-qos_mod: Any = sys.modules.get("rclpy.qos")
-if qos_mod is None:
-    qos_mod = _import_module_if_available("rclpy.qos")
-if qos_mod is None:
-    qos_mod = types.ModuleType("rclpy.qos")
-    sys.modules["rclpy.qos"] = qos_mod
-if not hasattr(qos_mod, "QoSProfile"):
-    qos_mod.QoSProfile = type("QoSProfile", (), {})
-if not hasattr(qos_mod, "QoSReliabilityPolicy"):
-    qos_mod.QoSReliabilityPolicy = type("QoSReliabilityPolicy", (), {})
-if not hasattr(qos_mod, "QoSHistoryPolicy"):
-    qos_mod.QoSHistoryPolicy = type("QoSHistoryPolicy", (), {})
-if not hasattr(qos_mod, "QoSDurabilityPolicy"):
-    qos_mod.QoSDurabilityPolicy = type("QoSDurabilityPolicy", (), {})
-
-rclpy.node = node_mod
-rclpy.executors = executors_mod
-rclpy.clock = clock_mod
-rclpy.parameter = parameter_mod
-rclpy.validate_namespace = validate_namespace_mod
-rclpy.validate_node_name = validate_node_name_mod
-rclpy.qos = qos_mod
+install_auto_launch_import_stubs(Path(__file__).resolve().parents[1])
 
 from uav.runtime.ModeManager import ModeManager  # noqa: E402
