+++ controls/sae_2025_ws/src/uav/test/test_payload_dlz_convex_hull_masking.py
@@ -4,7 +4,6 @@ import importlib
 import sys
 from pathlib import Path
 from types import SimpleNamespace
-import types
 from typing import Any, cast
 
 import numpy as np
@@ -16,6 +15,7 @@ if str(PACKAGE_ROOT) not in sys.path:
     sys.path.insert(0, str(PACKAGE_ROOT))
 
 from uav.cv.dlz_convex_hull import build_dlz_hull_mask  # noqa: E402
+from uav.test_support.ros_stubs import install_payload_mode_import_stubs  # noqa: E402
 
 
 def _bgr_from_hsv(h: int, s: int, v: int) -> tuple[int, int, int]:
@@ -38,74 +38,7 @@ def _roi_ratio(mask: np.ndarray, x0: int, y0: int, x1: int, y1: int) -> float:
     return float(np.count_nonzero(region)) / float(region.size)
 
 
-def _install_import_stubs() -> None:
-    if "rclpy" not in sys.modules:
-        rclpy: Any = types.ModuleType("rclpy")
-        node_module: Any = types.ModuleType("rclpy.node")
-
-        class Node:
-            pass
-
-        node_module.Node = Node
-        rclpy.node = node_module
-        sys.modules.update({"rclpy": rclpy, "rclpy.node": node_module})
-
-    if "cv_bridge" not in sys.modules:
-        cv_bridge: Any = types.ModuleType("cv_bridge")
-
-        class CvBridge:
-            def imgmsg_to_cv2(self, *_args, **_kwargs):
-                raise NotImplementedError
-
-            def cv2_to_compressed_imgmsg(self, *_args, **_kwargs):
-                return SimpleNamespace(header=SimpleNamespace(stamp=None))
-
-        cv_bridge.CvBridge = CvBridge
-        sys.modules["cv_bridge"] = cv_bridge
-
-    if "sensor_msgs" not in sys.modules:
-        sensor_msgs: Any = types.ModuleType("sensor_msgs")
-        sensor_msgs_msg: Any = types.ModuleType("sensor_msgs.msg")
-        sensor_msgs_msg.CompressedImage = type("CompressedImage", (), {})
-        sensor_msgs_msg.Image = type("Image", (), {})
-        sensor_msgs.msg = sensor_msgs_msg
-        sys.modules.update(
-            {"sensor_msgs": sensor_msgs, "sensor_msgs.msg": sensor_msgs_msg}
-        )
-
-    if "uav.vehicles.Payload" not in sys.modules:
-        payload_module: Any = types.ModuleType("uav.vehicles.Payload")
-        payload_module.Payload = type("Payload", (), {})
-        sys.modules["uav.vehicles.Payload"] = payload_module
-
-    if "uav.vision_nodes" not in sys.modules:
-        vision_nodes: Any = types.ModuleType("uav.vision_nodes")
-        vision_nodes.PayloadAprilTagNode = type("PayloadAprilTagNode", (), {})
-        sys.modules["uav.vision_nodes"] = vision_nodes
-
-    if "uav.vision_nodes.payload_perception_common" not in sys.modules:
-        common: Any = types.ModuleType("uav.vision_nodes.payload_perception_common")
-        common.DEFAULT_TAG_FAMILY = "tag36h11"
-        sys.modules["uav.vision_nodes.payload_perception_common"] = common
-
-    if "uav_interfaces" not in sys.modules:
-        sys.modules["uav_interfaces"] = types.ModuleType("uav_interfaces")
-
-    if "uav_interfaces.srv" not in sys.modules:
-        srv_module: Any = types.ModuleType("uav_interfaces.srv")
-
-        class PayloadAprilTagState:
-            class Request:
-                pass
-
-            class Response:
-                pass
-
-        srv_module.PayloadAprilTagState = PayloadAprilTagState
-        sys.modules["uav_interfaces.srv"] = srv_module
-
-
-_install_import_stubs()
+install_payload_mode_import_stubs()
 
 PayloadCornerNavigateMode = importlib.import_module(
     "uav.modes.payload.PayloadCornerNavigateMode"
